+++ R/print.R
@@ -12,8 +12,7 @@
 print.gn <- function(x, ...){
   stopifnot(inherits(x, "gn"))
   cat(paste("Geometric network in", x$q, "dimensions with",
-            x$W, "vertices and", x$M, "line segments.\n"))
-  cat(paste("Length of the network:", round(sum(x$d), 3), "units.\n"))
+            x$W, "vertices and", x$M, "curve segments.\n"))
   invisible(NULL)
 }
 
@@ -25,37 +24,26 @@ print.gn <- function(x, ...){
 print.gnpp <- function(x, ...){
   stopifnot(inherits(x, "gnpp"))
   cat(paste0("Point pattern of size ", nrow(x$data),
-            " on a geometric network in ",
-            x$network$q, " dimensions.\n",
-            "Network has ", x$network$W, " vertices and ",
-            x$network$M, " curve segments.\n"))
-  cat(paste("Length of the network:", round(sum(x$d), 3), "units.\n"))
-  if (ncol(x$data) > 6) {
-    n <- ncol(x$data) - 6
-    if (n == 1){
-      cat("Point hattern has one covariate:\n")
-    } else {
-      cat(paste0("Point pattern has ", n, " covariates:\n"))
-    }
-
-    for (i in 1:n) {
-      covariate <- x$data[[i + 6]]
-      cl <- as.character(class(covariate))
-      cat(paste0(i, ") '", names(x$data[i + 6]), "': Covariate of class ",
-                 cl, " "))
-      if (cl == "numeric") {
-        cat(paste0("with ",  length(unique(covariate)), " unique values.\n"))
-      }
-      if (cl == "factor") {
-        cat(paste0("with ", nlevels(covariate), " levels.\n"))
-      }
-      summary_cov <- t(as.matrix(summary(x$data %>% pull(7))))
-      rownames(summary_cov) <- ""
-      printCoefmat(summary_cov)
-      cat("\n")
-    }
+             " on a geometric network in ",
+             x$network$q, " dimensions.\n",
+             "Network has ", x$network$W, " vertices and ",
+             x$network$M, " curve segments.\n"))
+  cat("Point pattern has ")
+  if (ncol(x$network$lins) == 11) {
+    cat("no internal covariates and ")
   } else {
-    cat("Point pattern has no covariates.")
+    if (ncol(x$network$lins) == 12) {
+      cat("1 internal covariate and ")
+    } else {
+      cat(paste(ncol(x$network$lins) - 11), "internal covariates and ")
+    }
+  }
+  if (ncol(x$data) == 6) {
+    cat("no external covariates.")
+  } else if (ncol(x$data) == 7) {
+    cat("1 external covariate.")
+  } else {
+    cat(paste(ncol(x$data) - 6, "external covariates."))
   }
   invisible(NULL)
 }
@@ -65,19 +53,75 @@ print.gnpp <- function(x, ...){
 
 print.gnppfit <- function(x, ...){
   stopifnot(inherits(x, "gnppfit"))
-  cat(paste("Geometric network in", x$network$q, "dimensions"))
+  cat(paste0("Fitted point prozess on a geometric network in ",
+             x$network$q, " dimensions.\n",
+             "Network has ", x$network$W, " vertices and ",
+             x$network$M, " curve segments.\n"))
   invisible(NULL)
 }
 
-#' Print Method for gnppfit Summary
+#' Print Method for Summaries
 #'
-#' Prints basic information on the model fit and the underlying geometric
-#' network.
+#' Prints basic information of a geonet summary object.
 #'
-#' @param x A model fit on a geometric network (object of class \code{gnpp}).
+#' @param x A geonet summary object of class \code{summary.gn},
+#' code{summary.gnpp} or \code{summary.gnppfit}.
 #' @param ... Other arguments.
-#' @importFrom stats printCoefmat
 #' @return Invisibly returns the NULL object.
+#' @export
+
+print.summary.gn <- function(x, ...) {
+  stopifnot(inherits(x, "summary.gn"))
+  cat(paste("Geometric network in", x$q, "dimensions with",
+            x$W, "vertices and", x$M, "curve segments.\n"))
+  cat(paste("Total length of the network:", round(x$total_length, 3),
+            x$unit$plural, "\n"))
+  cat(paste("Minimum segment length:", round(x$range_length[1], 3),
+            x$unit$plural, "\n"))
+  cat(paste("Maximum segment length:", round(x$range_length[2], 3),
+            x$unit$plural, "\n"))
+  cat(paste("Distribution of vertex degrees:"))
+  print(x$degrees)
+}
+
+#' @rdname summary.gn
+#' @export
+
+print.summary.gnpp <- function(x, ...) {
+  stopifnot(inherits(x, "summary.gnpp"))
+  cat(paste("Point pattern on a geometric network in", x$q, "dimensions with",
+            x$W, "vertices and", x$M, "curve segments.\n"))
+  cat(paste("Total length of the network:", round(x$total_length, 3),
+            x$unit$plural, "\n"))
+  if (length(x$covariates$internal) > 0) {
+    cat(paste("Number of network internal covariates:",
+              length(x$covariates$internal), "\n"))
+    for (i in 1:length(x$covariates$internal)) {
+      cat(paste(i, ") ", x$covariates$internal[[i]]$class, " variable \"",
+                names(x$covariates$internal)[i], "\":\n", sep = ""))
+      print(x$covariates$internal[[i]]$summary)
+    }
+  } else {
+    cat("Network has no internal covariates\n")
+  }
+  cat(paste("Number of points:", x$n_points, "\n"))
+  cat(paste("Average intensity:", round(x$average_intensity, 5), "per",
+            x$unit$singular, "\n"))
+  if (length(x$covariates$internal) > 0) {
+    cat(paste("Number of external covariates:",
+              length(x$covariates$external), "\n"))
+    for (j in (i+1):(length(x$covariates$external)+i)) {
+      cat(paste(j, ") ", x$covariates$external[[j-i]]$class, " variable \"",
+                names(x$covariates$external)[j-i], "\":\n", sep = ""))
+      print(x$covariates$external[[j-i]]$summary)
+    }
+  } else {
+    cat("Point pattern has no external covariates\n")
+  }
+}
+
+#' @rdname summary.gn
+#' @importFrom stats printCoefmat
 #' @export
 
 print.summary.gnppfit <- function(x, ...){