+++ NEWS.md
@@ -2,6 +2,11 @@ geonet version 0.3.0 (Github release)
 
 - algorithm options can now be supplied to intensity_pspline via the
   "control" argument
+- as_gn now also takes the "units" attribute from a linnet object  
+- added missing summary and print methods
+- fixed bug in "internal" occurring when a linear covariates has length 1
+- internal covariates are recognized automatically and do not need to be 
+  assigned in the formula
 
 geonet version 0.2.0 (GitHub release)